[examples] reorganize code examples for BiDi
