--- examples/dotnet/SeleniumDocs/Bidirectional/WebDriverBiDi/BrowsingContextTest.cs
@@ -1,10 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace SeleniumDocs.Bidirectional.WebDriverBiDi
-{
-    [TestClass]
-    public class BrowsingContextTest : BaseTest
-    {
-
-    }
-}
