--- examples/dotnet/SeleniumDocs/Bidirectional/WebDriverBiDi/LogTest.cs
@@ -1,10 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace SeleniumDocs.Bidirectional.WebDriverBiDi
-{
-    [TestClass]
-    public class LogTest : BaseTest
-    {
-
-    }
-}
