+++ examples/java/src/test/java/dev/selenium/bidi/chrome_devtools/CdpEndpointTest.java
