+++ examples/java/src/test/java/dev/selenium/bidi/webdriver_bidi/LogTest.java
