+++ examples/python/tests/bidi/chrome_devtools/test_script.py
@@ -1,0 +1,15 @@
+import pytest
+from selenium.webdriver.common.by import By
+from selenium.webdriver.common.log import Log
+
+
+@pytest.mark.trio
+async def test_mutation(driver):
+    async with driver.bidi_connection() as session:
+        log = Log(driver, session)
+
+        async with log.mutation_events() as event:
+            driver.get('https://www.selenium.dev/selenium/web/dynamic.html')
+            driver.find_element(By.ID, "reveal").click()
+
+    assert event["element"] == driver.find_element(By.ID, "revealed")
