--- examples/python/tests/bidirectional/webdriver_bidi/test_browsing_context.py
@@ -1,2 +1,0 @@
-from selenium import webdriver
-
