+++ examples/python/tests/conftest.py
@@ -13,8 +13,18 @@ from selenium import webdriver
 
 
 @pytest.fixture(scope='function')
-def driver():
-    driver = webdriver.Chrome()
+def driver(request):
+    marker = request.node.get_closest_marker("driver_type")
+    driver_type = marker.args[0] if marker else None
+
+    if driver_type == "bidi":
+        options = webdriver.ChromeOptions()
+        options.enable_bidi = True
+        driver = webdriver.Chrome(options=options)
+    elif driver_type == "firefox":
+        driver = webdriver.Firefox()
+    else:
+        driver = webdriver.Chrome()
 
     yield driver
 
@@ -23,7 +33,7 @@ def driver():
 
 @pytest.fixture(scope='function')
 def chromedriver_bin():
-    service = webdriver.chrome.service.Service()
+    service = webdriver.ChromeService()
     options = webdriver.ChromeOptions()
     options.browser_version = 'stable'
     yield webdriver.common.driver_finder.DriverFinder(service=service, options=options).get_driver_path()
@@ -31,7 +41,7 @@ def chromedriver_bin():
 
 @pytest.fixture(scope='function')
 def chrome_bin():
-    service = webdriver.chrome.service.Service()
+    service = webdriver.ChromeService()
     options = webdriver.ChromeOptions()
     options.browser_version = 'stable'
     yield webdriver.common.driver_finder.DriverFinder(service=service, options=options).get_browser_path()
@@ -39,7 +49,7 @@ def chrome_bin():
 
 @pytest.fixture(scope='function')
 def edge_bin():
-    service = webdriver.edge.service.Service()
+    service = webdriver.EdgeService()
     options = webdriver.EdgeOptions()
     options.browser_version = 'stable'
     yield webdriver.common.driver_finder.DriverFinder(service=service, options=options).get_browser_path()
@@ -47,7 +57,7 @@ def edge_bin():
 
 @pytest.fixture(scope='function')
 def firefox_bin():
-    service = webdriver.firefox.service.Service()
+    service = webdriver.FirefoxService()
     options = webdriver.FirefoxOptions()
     options.browser_version = 'stable'
     yield webdriver.common.driver_finder.DriverFinder(service=service, options=options).get_browser_path()
