--- examples/ruby/spec/bidirectional/webdriver_bidi/log_spec.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-RSpec.describe 'Log Inspector' do
-  it 'listens to console log'
-  it 'gets JS exceptions'
-  it 'gets JS logs'
-end
